HttpRequest 파싱해주는 메서드들은 Http11Parser클래스로 따로 분리
